fix: break circular dependency between tool-manifest.ts and skills/load.ts
Move registerExternalTools/getExternalTools from tool-manifest.ts to
registry.ts so that meet-join/register.ts no longer imports
tool-manifest.ts. This breaks the circular module dependency chain:

  skills/load.ts → catalog-install.ts → capability-seed.ts →
  cli/program.ts → default-action.ts → lifecycle.ts →
  providers-setup.ts → external-skills-bootstrap.ts →
  meet-join/register.ts → tool-manifest.ts → skills/load.ts

The cycle caused a ReferenceError (TDZ) when any test imported
skills/load.ts, since tool-manifest.ts tried to access skillLoadTool
before skills/load.ts finished initializing.
